Prepare new upstream release.
The  previous tagged v0.2.6 had an issue which brokes the tests.
The tag was quickly deleted but it seems sum.golang.org has seen it
already and now complains that the download does NOT match the one
reported by the checksum server.

To not block users from upgrading a v0.2.7 will be released.